修复Data构造函数的bug，参数seq应该为ULL，之前使用int数字过大（2*10^9）就会出错了

+++ src/lvmisc.h
@@ -46,7 +46,7 @@ struct Data{
     ULL _sequenceNumber;
     int _op;
     Data()= default;
-    Data(int seq,int op,string k,string v=""):_sequenceNumber(seq),_op(op),_key(k),_value(v){}
+    Data(ULL seq,int op,string k,string v=""):_sequenceNumber(seq),_op(op),_key(k),_value(v){}
     friend ostream& operator<<(ostream& os,Data& a){
         return os<<setiosflags(ios::left)<<setw(4)<<a._key<<" : "<<a._value<<"\t"<<a._op<<"\t"<<a._sequenceNumber;
     }
